dma-buf/sync_file: Speed up ioctl by omitting debug names
A lot of CPU time is wasted on allocating, populating, and copying
debug names back and forth with userspace when they're not actually
needed. We can't just remove the name buffers from the various sync data
structures though because we must preserve ABI compatibility with
userspace, but instead we can just pretend the name fields of the
user-shared structs aren't there. This massively reduces the sizes of
memory allocated for these data structures and the amount of data passed
between userspace, as well as eliminates a kzalloc() entirely from
sync_file_ioctl_fence_info(), thus improving graphics performance.

Change-Id: I8a5d6b401087df13c12717f25bad0a636de4a7bf
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[fiqri19102002: Keep user_name field for Android KABI compatibility]
Signed-off-by: Fiqri Ardyansyah <fiqri191002@gmail.com>